4554 client - move Skills Copilot trigger into Create Skill dropdown
Restore the Create With AI item in AiSkillsCreateDropdown (ff-4554) and
wire its click to setContext(Source.SKILLS, MODE.BUILD) +
setCopilotPanelOpen(true) so it opens the global side Copilot panel
instead of navigating to a dedicated page. Drop the standalone sparkles
button (and its handler + imports) from the Skills header.
